+++ tests/options/mod.rs
@@ -1,2 +1,1 @@
 mod exclude;
-mod recursive;